requirements.txt
torch
torchvision
pillow
numpy
scikit-learn
requests